update todo and break out into new file
