+++ package/parser/definition.go
@@ -64,26 +64,3 @@ func (pkg *Package) definition(name string) (decl Declaration, err error) {
 	})
 	return decl, err
 }
-
-// builtin declaration
-type builtin string
-
-// Name is the built-in type
-func (b builtin) Name() string {
-	return string(b)
-}
-
-func (b builtin) Kind() Kind {
-	return KindBuiltin
-}
-
-// Directory for builtin is blank
-func (b builtin) Directory() string {
-	return ""
-}
-
-// Package for builtin is blank
-// TODO: there should probably be a built-in package
-func (b builtin) Package() *Package {
-	return nil
-}
